More experiment setups, more collider types.
Bug fix.
Added log experiment.
Improved box selection and visualization.
