+++ EvoEngine_Plugins/DigitalAgriculture/include/SorghumModel/SorghumDescriptor.hpp
@@ -6,7 +6,7 @@ struct SorghumMeshGeneratorSettings {
   bool enable_panicle = true;
   bool enable_stem = true;
   bool enable_leaves = true;
-  bool enable_leaf_sheath = false;
+  bool enable_leaf_sheath = true;
   int single_leaf_index = -1;
   bool bottom_face = true;
   bool leaf_separated = false;
