+++ EvoEngine_Plugins/EcoSysLab/include/DynamicTreeStrands.hpp
@@ -25,17 +25,12 @@ class DynamicTreeStrands : public IPrivateComponent {
     Entity target_entity;
     std::shared_ptr<DsTransform> ds_transform;
   };
-  struct EntityAttraction {
-    Entity target_entity;
-    std::shared_ptr<DsAttraction> ds_attraction;
-  };
   
   bool limit_strand_length = true;
   float max_strand_length = 1.f;
 
   std::vector<EntityTransform> transform_operators;
-  std::vector<EntityAttraction> attraction_operators;
-
+  
   std::shared_ptr<DsBoxSelection> box_selection_operator;
   std::shared_ptr<DsDrag> drag_operator;
   std::shared_ptr<DsGravity> gravity;
@@ -47,16 +42,27 @@ class DynamicTreeStrands : public IPrivateComponent {
   void OnDestroy() override;
   void CollectAssetRef(std::vector<AssetRef>& list) override;
 
-  struct MultipleRodExperimentSetupSettings {
+  struct BoardExperimentSetupSettings {
     float segment_length = 0.05f;
     float radius = 0.002f;
-    glm::ivec3 rod_dimension = {10, 10, 10};
+    glm::ivec3 rod_dimension = {10, 40, 20};
     bool add_operator = true;
   };
-
-  void MultipleRodExperimentSetup(const MultipleRodExperimentSetupSettings& settings);
+  struct LogExperimentSetupSettings {
+    float segment_length = 0.05f;
+    float radius = 0.002f;
+    int rod_size = 400;
+    int rod_segment_count = 10;
+    float center_attraction_strength = 40000;
+    bool add_operator = true;
+  };
+  void BoardExperimentSetup(const BoardExperimentSetupSettings& settings);
+  void LogExperimentSetup(const LogExperimentSetupSettings& settings);
   void InitializeStrandParticles(const DtsStrandGroup& target_strand_group) const;
   void ClearStrandParticles() const;
+
+  void InteractionStep() const;
+
   void PhysicsStep() const;
 
   void Visualization(const std::shared_ptr<Camera>& target_camera) const;
