+++ EvoEngine_Plugins/EcoSysLab/src/DsColliders.cpp
@@ -4,22 +4,11 @@
 
 using namespace eco_sys_lab_plugin;
 
-void DsBoxCollider::RenderBound(const std::shared_ptr<EditorLayer>& editor_layer, const std::shared_ptr<Camera>& editor_camera, const glm::vec4& color) {
-  const auto mesh_renderer = mesh_renderer_ref.Get<MeshRenderer>();
-  if (!mesh_renderer)
-    return;
-  if (!mesh_renderer->IsEnabled())
-    return;
-  const auto mesh = mesh_renderer->mesh.Get<Mesh>();
-  const auto material = mesh_renderer->material.Get<Material>();
-  if (!mesh || !material)
-    return;
-
-  const auto& bound = mesh->GetBound();
-
+void DsBoxCollider::RenderBound(const std::shared_ptr<EditorLayer>& editor_layer,
+                                const std::shared_ptr<Camera>& editor_camera, const glm::vec4& color) {
   const auto scene = GetScene();
   const auto global_transform = scene->GetDataComponent<GlobalTransform>(GetOwner());
-  glm::vec3 size = bound.Size() * 2.0f;
+  glm::vec3 size = scale * global_transform.GetScale() * 2.0f;
   if (size.x < 0.001f)
     size.x = 0.001f;
   if (size.z < 0.001f)
@@ -33,17 +22,18 @@ void DsBoxCollider::RenderBound(const std::shared_ptr<EditorLayer>& editor_layer
   gizmo_settings.draw_settings.line_width = 1.0f;
   gizmo_settings.depth_test = true;
   editor_layer->DrawGizmoMesh(Resources::GetResource<Mesh>("PRIMITIVE_CUBE"), editor_camera, color,
-                              glm::translate(bound.Center()) * glm::scale(size) * global_transform.value, 1,
-                              gizmo_settings);
+                              glm::translate(global_transform.GetPosition()) *
+                                  glm::mat4_cast(global_transform.GetRotation()) * glm::scale(size),
+                              1, gizmo_settings);
 }
 
 DsBoxCollider::DsBoxCollider() {
   if (!pipeline) {
     static std::shared_ptr<Shader> shader{};
     shader = std::make_shared<Shader>();
-    shader->Set(
-        ShaderType::Compute, Platform::Constants::shader_global_defines,
-        std::filesystem::path("./EcoSysLabResources") / "Shaders/Compute/DynamicStrands/ContactConstraints/BoxCollider.comp");
+    shader->Set(ShaderType::Compute, Platform::Constants::shader_global_defines,
+                std::filesystem::path("./EcoSysLabResources") /
+                    "Shaders/Compute/DynamicStrands/ContactConstraints/BoxCollider.comp");
 
     pipeline = std::make_shared<ComputePipeline>();
     pipeline->compute_shader = shader;
@@ -60,7 +50,18 @@ DsBoxCollider::DsBoxCollider() {
 
 bool DsBoxCollider::OnInspect(const std::shared_ptr<EditorLayer>& editor_layer) {
   bool changed = false;
-  if (editor_layer->DragAndDropButton<MeshRenderer>(mesh_renderer_ref, "Mesh Renderer")) {
+  static PrivateComponentRef mesh_renderer_ref;
+  if (editor_layer->DragAndDropButton<MeshRenderer>(mesh_renderer_ref, "Apply bound from Mesh Renderer")) {
+    if (const auto mesh_renderer = mesh_renderer_ref.Get<MeshRenderer>()) {
+      if (const auto mesh = mesh_renderer->mesh.Get<Mesh>()) {
+        const auto& bound = mesh->GetBound();
+        scale = bound.Size();
+      }
+    }
+    mesh_renderer_ref.Clear();
+  }
+
+  if (ImGui::DragFloat3("Scale", &scale.x, 0.01f, 0.0f, 10.0f)) {
     changed = true;
   }
 
@@ -68,38 +69,24 @@ bool DsBoxCollider::OnInspect(const std::shared_ptr<EditorLayer>& editor_layer) 
     changed = true;
   }
 
+  ImGui::ColorEdit4("Bound Color:##DsBoxCollider", (float*)(void*)&bound_color);
   static bool display_bound = true;
   ImGui::Checkbox("Display bounds##DsBoxCollider", &display_bound);
   if (display_bound) {
-    static auto display_bound_color = glm::vec4(1.0f, 0.0f, 1.0f, 0.2f);
-    ImGui::ColorEdit4("Color:##DsBoxCollider", (float*)(void*)&display_bound_color);
-    RenderBound(editor_layer, editor_layer->GetSceneCamera(), display_bound_color);
+    RenderBound(editor_layer, editor_layer->GetSceneCamera(), bound_color);
   }
-
   return changed;
 }
 
 void DsBoxCollider::ProjectPositionConstraint(const DynamicStrands::PhysicsParameters& physics_parameters,
-    const DynamicStrands& target_dynamic_strands) {
-  const auto mesh_renderer = mesh_renderer_ref.Get<MeshRenderer>();
-  if (!mesh_renderer)
-    return;
-  if (!mesh_renderer->IsEnabled())
-    return;
-  const auto mesh = mesh_renderer->mesh.Get<Mesh>();
-  const auto material = mesh_renderer->material.Get<Material>();
-  if (!mesh || !material)
-    return;
-
-  const auto& bound = mesh->GetBound();
-
+                                              const DynamicStrands& target_dynamic_strands) {
   const auto scene = GetScene();
   const auto global_transform = scene->GetDataComponent<GlobalTransform>(GetOwner());
 
   const auto current_frame_index = Platform::GetCurrentFrameIndex();
   const uint32_t task_work_group_invocations =
       Platform::GetSelectedPhysicalDevice()->mesh_shader_properties_ext.maxPreferredTaskWorkGroupInvocations;
-  glm::vec3 size = bound.Size();
+  glm::vec3 size = scale * global_transform.GetScale();
   if (size.x < 0.001f)
     size.x = 0.001f;
   if (size.z < 0.001f)
@@ -107,8 +94,8 @@ void DsBoxCollider::ProjectPositionConstraint(const DynamicStrands::PhysicsParam
   if (size.y < 0.001f)
     size.y = 0.001f;
   PushConstant segment_push_constant;
-  segment_push_constant.obb_center = global_transform.GetPosition() + bound.Center();
-  segment_push_constant.obb_scale = size * global_transform.GetScale();
+  segment_push_constant.obb_center = global_transform.GetPosition();
+  segment_push_constant.obb_scale = size;
   segment_push_constant.obb_rotation = global_transform.GetRotation();
   segment_push_constant.segment_size = target_dynamic_strands.segments.size();
   segment_push_constant.softness = softness;
@@ -126,21 +113,241 @@ void DsBoxCollider::ProjectPositionConstraint(const DynamicStrands::PhysicsParam
   });
 }
 
-
-void DsBoxCollider::OnDestroy() {
-  mesh_renderer_ref.Clear();
-}
-
 void DsBoxCollider::Serialize(YAML::Emitter& out) const {
-  mesh_renderer_ref.Save("mesh_renderer_ref", out);
+  out << YAML::Key << "bound_color" << YAML::Value << bound_color;
+  out << YAML::Key << "scale" << YAML::Value << scale;
+  out << YAML::Key << "softness" << YAML::Value << softness;
 }
 
 void DsBoxCollider::Deserialize(const YAML::Node& in) {
-  mesh_renderer_ref.Load("mesh_renderer_ref", in, GetScene());
+  if (in["bound_color"])
+    bound_color = in["bound_color"].as<glm::vec4>();
+  if (in["scale"])
+    scale = in["scale"].as<glm::vec3>();
+  if (in["softness"])
+    softness = in["softness"].as<float>();
 }
 
-void DsBoxCollider::Relink(const std::unordered_map<Handle, Handle>& map, const std::shared_ptr<Scene>& scene) {
-  mesh_renderer_ref.Relink(map, scene);
+void DsCylinderCollider::RenderBound(const std::shared_ptr<EditorLayer>& editor_layer,
+                                     const std::shared_ptr<Camera>& editor_camera, const glm::vec4& color) {
+  const auto scene = GetScene();
+  const auto global_transform = scene->GetDataComponent<GlobalTransform>(GetOwner());
+  const auto scale = global_transform.GetScale();
+  auto size = glm::vec2(radius * glm::max(scale.x, scale.z), height * scale.y) * 2.f;
+  if (size.x < 0.001f)
+    size.x = 0.001f;
+  if (size.y < 0.001f)
+    size.y = 0.001f;
+  GizmoSettings gizmo_settings;
+  gizmo_settings.draw_settings.cull_mode = VK_CULL_MODE_NONE;
+  gizmo_settings.draw_settings.blending = true;
+  gizmo_settings.draw_settings.polygon_mode = VK_POLYGON_MODE_FILL;
+  gizmo_settings.draw_settings.line_width = 1.0f;
+  gizmo_settings.depth_test = true;
+  editor_layer->DrawGizmoMesh(Resources::GetResource<Mesh>("PRIMITIVE_CYLINDER"), editor_camera, color,
+                              glm::translate(global_transform.GetPosition()) *
+                                  glm::mat4_cast(global_transform.GetRotation()) *
+                                  glm::scale(glm::vec3(size.x, size.y, size.x)),
+                              1, gizmo_settings);
 }
 
+DsCylinderCollider::DsCylinderCollider() {
+  if (!pipeline) {
+    static std::shared_ptr<Shader> shader{};
+    shader = std::make_shared<Shader>();
+    shader->Set(ShaderType::Compute, Platform::Constants::shader_global_defines,
+                std::filesystem::path("./EcoSysLabResources") /
+                    "Shaders/Compute/DynamicStrands/ContactConstraints/CylinderCollider.comp");
 
+    pipeline = std::make_shared<ComputePipeline>();
+    pipeline->compute_shader = shader;
+    pipeline->descriptor_set_layouts.emplace_back(DynamicStrands::strands_layout);
+
+    auto& push_constant_range = pipeline->push_constant_ranges.emplace_back();
+    push_constant_range.size = sizeof(PushConstant);
+    push_constant_range.offset = 0;
+    push_constant_range.stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
+
+    pipeline->Initialize();
+  }
+}
+
+bool DsCylinderCollider::OnInspect(const std::shared_ptr<EditorLayer>& editor_layer) {
+  bool changed = false;
+
+  if (ImGui::DragFloat("Radius", &radius, 0.01f, 0.0f, 10.0f)) {
+    changed = true;
+  }
+  if (ImGui::DragFloat("Height", &height, 0.01f, 0.0f, 10.0f)) {
+    changed = true;
+  }
+  if (ImGui::DragFloat("Softness", &softness, 0.01f, 0.0f, 1.0f)) {
+    changed = true;
+  }
+  ImGui::ColorEdit4("Bound Color:##DsBoxCollider", (float*)(void*)&bound_color);
+  static bool display_bound = true;
+  ImGui::Checkbox("Display bounds##DsBoxCollider", &display_bound);
+  if (display_bound) {
+    RenderBound(editor_layer, editor_layer->GetSceneCamera(), bound_color);
+  }
+  return changed;
+}
+
+void DsCylinderCollider::ProjectPositionConstraint(const DynamicStrands::PhysicsParameters& physics_parameters,
+                                                   const DynamicStrands& target_dynamic_strands) {
+  const auto scene = GetScene();
+  const auto global_transform = scene->GetDataComponent<GlobalTransform>(GetOwner());
+
+  const auto current_frame_index = Platform::GetCurrentFrameIndex();
+  const uint32_t task_work_group_invocations =
+      Platform::GetSelectedPhysicalDevice()->mesh_shader_properties_ext.maxPreferredTaskWorkGroupInvocations;
+  const glm::vec3 scale = global_transform.GetScale();
+  auto size = glm::vec2(radius * glm::max(scale.x, scale.z), height * scale.y);
+  if (size.x < 0.001f)
+    size.x = 0.001f;
+  if (size.y < 0.001f)
+    size.y = 0.001f;
+  PushConstant segment_push_constant;
+  segment_push_constant.obb_center = global_transform.GetPosition();
+  segment_push_constant.radius = size.x;
+  segment_push_constant.height = size.y;
+  segment_push_constant.obb_rotation = global_transform.GetRotation();
+  segment_push_constant.segment_size = target_dynamic_strands.segments.size();
+  segment_push_constant.softness = softness;
+
+  Platform::RecordCommandsMainQueue([&](const VkCommandBuffer vk_command_buffer) {
+    pipeline->Bind(vk_command_buffer);
+    pipeline->BindDescriptorSet(
+        vk_command_buffer, 0,
+        target_dynamic_strands.strands_descriptor_sets[current_frame_index]->GetVkDescriptorSet());
+
+    pipeline->PushConstant(vk_command_buffer, 0, segment_push_constant);
+    vkCmdDispatch(vk_command_buffer, Platform::DivUp(segment_push_constant.segment_size, task_work_group_invocations),
+                  1, 1);
+    Platform::EverythingBarrier(vk_command_buffer);
+  });
+}
+
+void DsCylinderCollider::Serialize(YAML::Emitter& out) const {
+  out << YAML::Key << "bound_color" << YAML::Value << bound_color;
+  out << YAML::Key << "radius" << YAML::Value << radius;
+  out << YAML::Key << "height" << YAML::Value << height;
+  out << YAML::Key << "softness" << YAML::Value << softness;
+}
+
+void DsCylinderCollider::Deserialize(const YAML::Node& in) {
+  if (in["bound_color"])
+    bound_color = in["bound_color"].as<glm::vec4>();
+  if (in["radius"])
+    radius = in["radius"].as<float>();
+  if (in["height"])
+    height = in["height"].as<float>();
+  if (in["softness"])
+    softness = in["softness"].as<float>();
+}
+
+void DsSphereCollider::RenderBound(const std::shared_ptr<EditorLayer>& editor_layer,
+                                   const std::shared_ptr<Camera>& editor_camera, const glm::vec4& color) {
+  const auto scene = GetScene();
+  const auto global_transform = scene->GetDataComponent<GlobalTransform>(GetOwner());
+  const auto scale = global_transform.GetScale();
+  GizmoSettings gizmo_settings;
+  gizmo_settings.draw_settings.cull_mode = VK_CULL_MODE_NONE;
+  gizmo_settings.draw_settings.blending = true;
+  gizmo_settings.draw_settings.polygon_mode = VK_POLYGON_MODE_FILL;
+  gizmo_settings.draw_settings.line_width = 1.0f;
+  gizmo_settings.depth_test = true;
+  editor_layer->DrawGizmoMesh(
+      Resources::GetResource<Mesh>("PRIMITIVE_SPHERE"), editor_camera, color,
+      glm::translate(global_transform.GetPosition()) *
+          glm::scale(glm::vec3(glm::max(0.001f, radius * 2.f)) * glm::max(glm::max(scale.x, scale.y), scale.z)),
+      1, gizmo_settings);
+}
+
+DsSphereCollider::DsSphereCollider() {
+  if (!pipeline) {
+    static std::shared_ptr<Shader> shader{};
+    shader = std::make_shared<Shader>();
+    shader->Set(ShaderType::Compute, Platform::Constants::shader_global_defines,
+                std::filesystem::path("./EcoSysLabResources") /
+                    "Shaders/Compute/DynamicStrands/ContactConstraints/SphereCollider.comp");
+
+    pipeline = std::make_shared<ComputePipeline>();
+    pipeline->compute_shader = shader;
+    pipeline->descriptor_set_layouts.emplace_back(DynamicStrands::strands_layout);
+
+    auto& push_constant_range = pipeline->push_constant_ranges.emplace_back();
+    push_constant_range.size = sizeof(PushConstant);
+    push_constant_range.offset = 0;
+    push_constant_range.stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
+
+    pipeline->Initialize();
+  }
+}
+
+bool DsSphereCollider::OnInspect(const std::shared_ptr<EditorLayer>& editor_layer) {
+  bool changed = false;
+  if (ImGui::DragFloat("Radius", &radius, 0.01f, 0.0f, 10.0f)) {
+    changed = true;
+  }
+
+  if (ImGui::DragFloat("Softness", &softness, 0.01f, 0.0f, 1.0f)) {
+    changed = true;
+  }
+
+  ImGui::ColorEdit4("Bound Color:##DsBoxCollider", (float*)(void*)&bound_color);
+  static bool display_bound = true;
+  ImGui::Checkbox("Display bounds##DsBoxCollider", &display_bound);
+  if (display_bound) {
+    RenderBound(editor_layer, editor_layer->GetSceneCamera(), bound_color);
+  }
+
+  return changed;
+}
+
+void DsSphereCollider::ProjectPositionConstraint(const DynamicStrands::PhysicsParameters& physics_parameters,
+                                                 const DynamicStrands& target_dynamic_strands) {
+  const auto scene = GetScene();
+  const auto global_transform = scene->GetDataComponent<GlobalTransform>(GetOwner());
+
+  const auto current_frame_index = Platform::GetCurrentFrameIndex();
+  const uint32_t task_work_group_invocations =
+      Platform::GetSelectedPhysicalDevice()->mesh_shader_properties_ext.maxPreferredTaskWorkGroupInvocations;
+  const glm::vec3 scale = global_transform.GetScale();
+  float size = radius * glm::max(glm::max(scale.x, scale.y), scale.z);
+  if (size < 0.001f)
+    size = 0.001f;
+
+  PushConstant segment_push_constant;
+  segment_push_constant.obb_center = global_transform.GetPosition();
+  segment_push_constant.radius = size;
+  segment_push_constant.segment_size = target_dynamic_strands.segments.size();
+  segment_push_constant.softness = softness;
+
+  Platform::RecordCommandsMainQueue([&](const VkCommandBuffer vk_command_buffer) {
+    pipeline->Bind(vk_command_buffer);
+    pipeline->BindDescriptorSet(
+        vk_command_buffer, 0,
+        target_dynamic_strands.strands_descriptor_sets[current_frame_index]->GetVkDescriptorSet());
+
+    pipeline->PushConstant(vk_command_buffer, 0, segment_push_constant);
+    vkCmdDispatch(vk_command_buffer, Platform::DivUp(segment_push_constant.segment_size, task_work_group_invocations),
+                  1, 1);
+    Platform::EverythingBarrier(vk_command_buffer);
+  });
+}
+
+void DsSphereCollider::Serialize(YAML::Emitter& out) const {
+  out << YAML::Key << "bound_color" << YAML::Value << bound_color;
+  out << YAML::Key << "softness" << YAML::Value << softness;
+  out << YAML::Key << "radius" << YAML::Value << radius;
+}
+
+void DsSphereCollider::Deserialize(const YAML::Node& in) {
+  if (in["bound_color"])
+    bound_color = in["bound_color"].as<glm::vec4>();
+  if (in["radius"])
+    radius = in["radius"].as<float>();
+  if (in["softness"])
+    softness = in["softness"].as<float>();
+}
