+++ EvoEngine_Plugins/EcoSysLab/src/EcoSysLabLayerSoils.cpp
@@ -1,0 +1,317 @@
+//
+// Created by lllll on 11/1/2022.
+//
+
+#include "EcoSysLabLayer.hpp"
+#ifdef OPTIX_RAY_TRACER_PLUGIN
+#  include <RayTracerLayer.hpp>
+#endif
+#include "BarkDescriptor.hpp"
+#include "Times.hpp"
+
+#include "BillboardCloudsConverter.hpp"
+#include "ClassRegistry.hpp"
+#include "Climate.hpp"
+#include "CubeVolume.hpp"
+#include "DsColliders.hpp"
+#include "DsOperators.hpp"
+#include "DynamicTreeStrands.hpp"
+#include "FlowerDescriptor.hpp"
+#include "FoliageDescriptor.hpp"
+#include "ForestDescriptor.hpp"
+#include "FruitDescriptor.hpp"
+#include "RenderLayer.hpp"
+#include "Shader.hpp"
+#include "Soil.hpp"
+#include "SpatialPlantDistributionSimulator.hpp"
+#include "StrandsRenderer.hpp"
+#include "Tree.hpp"
+#include "TreeStructor.hpp"
+using namespace eco_sys_lab_plugin;
+
+AssetRegistration<HeightField> height_field_registry("HeightField", {".heightfield"});
+AssetRegistration<SoilLayerDescriptor> soil_layer_d_registry("SoilLayerDescriptor", {".soillayer"});
+
+PrivateComponentRegistration<Soil> soil_registry("Soil");
+
+AssetRegistration<ProceduralNoise2D> procedural_noise2d_registry("ProceduralNoise2D", {".noise2D"});
+AssetRegistration<ProceduralNoise3D> procedural_noise3d_registry("ProceduralNoise3D", {".noise3D"});
+AssetRegistration<SoilDescriptor> soil_d_registry("SoilDescriptor", {".soil"});
+void EcoSysLabLayer::OnSoilVisualizationMenu() {
+  static bool force_update;
+  ImGui::Checkbox("Force Update", &force_update);
+
+  if (ImGui::Checkbox("Vector Visualization", &vector_enable_)) {
+    if (vector_enable_)
+      update_vector_matrices_ = true;
+  }
+
+  if (ImGui::Checkbox("Scalar Visualization", &scalar_enable_)) {
+    if (scalar_enable_)
+      update_scalar_matrices_ = true;
+  }
+
+  if (vector_enable_) {
+    update_vector_matrices_ = update_vector_matrices_ || force_update;
+
+    if (ImGui::TreeNodeEx("Vector", ImGuiTreeNodeFlags_DefaultOpen)) {
+      if (ImGui::Button("Reset")) {
+        vector_multiplier_ = 50.0f;
+        vector_base_color_ = glm::vec4(1.0f, 1.0f, 1.0f, 0.8f);
+        vector_soil_property_ = 4;
+        vector_line_width_factor_ = 0.1f;
+        vector_line_max_width_ = 0.1f;
+        update_vector_matrices_ = true;
+      }
+      if (ImGui::ColorEdit4("Vector Base Color", &vector_base_color_.x)) {
+        update_vector_matrices_ = true;
+      }
+      if (ImGui::DragFloat("Multiplier", &vector_multiplier_, 0.1f, 0.0f, 100.0f, "%.3f")) {
+        update_vector_matrices_ = true;
+      }
+      if (ImGui::DragFloat("Line Width Factor", &vector_line_width_factor_, 0.01f, 0.0f, 5.0f)) {
+        update_vector_matrices_ = true;
+      }
+      if (ImGui::DragFloat("Max Line Width", &vector_line_max_width_, 0.01f, 0.0f, 5.0f)) {
+        update_vector_matrices_ = true;
+      }
+      if (ImGui::Combo("Vector Mode",
+                       {"N/A", "N/A", "Water Density Gradient", "Flux", "Divergence", "N/A", "N/A", "N/A"},
+                       vector_soil_property_)) {
+        update_vector_matrices_ = true;
+      }
+      ImGui::TreePop();
+    }
+  }
+  if (scalar_enable_) {
+    update_scalar_matrices_ = update_scalar_matrices_ || force_update;
+
+    if (scalar_enable_ && ImGui::TreeNodeEx("Scalar", ImGuiTreeNodeFlags_DefaultOpen)) {
+      if (ImGui::Button("Reset")) {
+        scalar_multiplier_ = 1.0f;
+        scalar_box_size_ = 0.5f;
+        scalar_min_alpha_ = 0.00f;
+        scalar_base_color_ = glm::vec3(0.0f, 0.0f, 1.0f);
+        scalar_soil_property_ = 1;
+        update_scalar_matrices_ = true;
+      }
+      if (ImGui::SliderFloat("X Depth", &soil_cutout_x_depth_, 0.0f, 1.0f)) {
+        update_scalar_matrices_ = true;
+      }
+      if (ImGui::SliderFloat("Z Depth", &soil_cutout_z_depth_, 0.0f, 1.0f)) {
+        update_scalar_matrices_ = true;
+      }
+
+      if (ImGui::TreeNodeEx("Layer colors", ImGuiTreeNodeFlags_DefaultOpen)) {
+        for (int i = 0; i < 10; i++) {
+          ImGui::ColorEdit4(("Layer " + std::to_string(i)).c_str(), &soil_layer_colors_[i].x);
+        }
+        ImGui::TreePop();
+      }
+
+      if (ImGui::ColorEdit3("Scalar Base Color", &scalar_base_color_.x)) {
+        update_scalar_matrices_ = true;
+      }
+      if (ImGui::SliderFloat("Multiplier", &scalar_multiplier_, 0.001, 10000, "%.4f", ImGuiSliderFlags_Logarithmic)) {
+        update_scalar_matrices_ = true;
+      }
+      if (ImGui::DragFloat("Min alpha", &scalar_min_alpha_, 0.001f, 0.0f, 1.0f)) {
+        update_scalar_matrices_ = true;
+      }
+      if (ImGui::DragFloat("Box size", &scalar_box_size_, 0.001f, 0.0f, 1.0f)) {
+        update_scalar_matrices_ = true;
+      }
+      // disable less useful visualizations to avoid clutter in the gui
+      if (ImGui::Combo(
+              "Scalar Mode",
+              {"Blank", "Water Density", "N/A", "N/A", "N/A", "Nutrient Density", "Soil Density", "Soil Layer"},
+              scalar_soil_property_)) {
+        update_scalar_matrices_ = true;
+      }
+      ImGui::TreePop();
+    }
+  }
+}
+void EcoSysLabLayer::SoilVisualization() {
+  std::shared_ptr<Soil> soil;
+  if (const auto soil_candidate = FindSoil(); !soil_candidate.expired())
+    soil = soil_candidate.lock();
+
+  if (!soil)
+    return;
+
+  const auto& soil_model = soil->soil_model;
+  if (soil_version_ != soil_model.m_version) {
+    update_vector_matrices_ = true;
+    update_scalar_matrices_ = true;
+    soil_version_ = soil_model.m_version;
+  }
+
+  if (vector_enable_) {
+    SoilVisualizationVector(soil_model);
+  }
+  if (scalar_enable_) {
+    SoilVisualizationScalar(soil_model);
+  }
+}
+
+void EcoSysLabLayer::SoilVisualizationScalar(const VoxelSoilModel& soil_model) {
+  const auto num_voxels = soil_model.m_resolution.x * soil_model.m_resolution.y * soil_model.m_resolution.z;
+  if (update_scalar_matrices_) {
+    std::vector<ParticleInfo> particle_infos;
+    particle_infos.resize(num_voxels);
+    Jobs::RunParallelFor(num_voxels, [&](unsigned i) {
+      const auto coordinate = soil_model.GetCoordinateFromIndex(i);
+      if (static_cast<float>(coordinate.x) / soil_model.m_resolution.x < soil_cutout_x_depth_ ||
+          static_cast<float>(coordinate.z) / soil_model.m_resolution.z > (1.0f - soil_cutout_z_depth_)) {
+        particle_infos[i].instance_matrix.value = glm::mat4(0.0f);
+      } else {
+        particle_infos[i].instance_matrix.value = glm::translate(soil_model.GetPositionFromCoordinate(coordinate)) *
+                                                  glm::mat4_cast(glm::quat(glm::vec3(0.0f))) *
+                                                  glm::scale(glm::vec3(soil_model.GetVoxelSize() * scalar_box_size_));
+      }
+    });
+    auto visualize_vec3 = [&](const Field& x, const Field& y, const Field& z) {
+      Jobs::RunParallelFor(num_voxels, [&](unsigned i) {
+        const auto value = glm::vec3(x[i], y[i], z[i]);
+        particle_infos[i].instance_color = {
+            glm::normalize(value), glm::clamp(glm::length(value) * scalar_multiplier_, scalar_min_alpha_, 1.0f)};
+      });
+    };
+
+    auto visualize_float = [&](const Field& v) {
+      Jobs::RunParallelFor(num_voxels, [&](unsigned i) {
+        const auto value = glm::vec3(v[i]);
+        particle_infos[i].instance_color = {
+            scalar_base_color_, glm::clamp(glm::length(value) * scalar_multiplier_, scalar_min_alpha_, 1.0f)};
+      });
+    };
+
+    switch (static_cast<SoilProperty>(scalar_soil_property_)) {
+      case SoilProperty::Blank: {
+        Jobs::RunParallelFor(num_voxels, [&](unsigned i) {
+          particle_infos[i].instance_color = {scalar_base_color_, 0.01f};
+        });
+      } break;
+      case SoilProperty::WaterDensity: {
+        visualize_float(soil_model.m_w);
+      } break;
+      case SoilProperty::NutrientDensity: {
+        visualize_float(soil_model.m_n);
+      } break;
+      case SoilProperty::SoilDensity: {
+        visualize_float(soil_model.m_d);
+      } break;
+      case SoilProperty::SoilLayer: {
+        Jobs::RunParallelFor(num_voxels, [&](unsigned i) {
+          const auto layerIndex = soil_model.m_material_id[i];
+          if (layerIndex == 0)
+            particle_infos[i].instance_color = glm::vec4(0.0f);
+          else {
+            particle_infos[i].instance_color = soil_layer_colors_[layerIndex - 1];
+          }
+        });
+      } break;
+        /*case SoilProperty::DiffusionDivergence:
+        {
+                visualize_vec3(soilModel.m_div_diff_x, soilModel.m_div_diff_y, soilModel.m_div_diff_z);
+        }break;*/
+      default: {
+        Jobs::RunParallelFor(num_voxels, [&](unsigned i) {
+          particle_infos[i].instance_color = {scalar_base_color_, 0.01f};
+        });
+      } break;
+    }
+    ground_fruit_matrices_->SetParticleInfos(particle_infos);
+  }
+  update_scalar_matrices_ = false;
+  const auto editor_layer = Application::GetLayer<EditorLayer>();
+  GizmoSettings gizmo_settings;
+  gizmo_settings.draw_settings.blending = true;
+  gizmo_settings.draw_settings.blending_src_factor = VK_BLEND_FACTOR_SRC_ALPHA;
+  gizmo_settings.draw_settings.blending_dst_factor = VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA;
+  gizmo_settings.draw_settings.cull_mode = VK_CULL_MODE_NONE;
+  editor_layer->DrawGizmoMeshInstancedColored(Resources::GetResource<Mesh>("PRIMITIVE_CUBE"), scalar_matrices_,
+                                              glm::mat4(1.0f), 1.0f, gizmo_settings);
+}
+
+void EcoSysLabLayer::SoilVisualizationVector(const VoxelSoilModel& soil_model) {
+  const auto num_voxels = soil_model.m_resolution.x * soil_model.m_resolution.y * soil_model.m_resolution.z;
+
+  if (update_vector_matrices_) {
+    std::vector<ParticleInfo> particle_infos;
+    particle_infos.resize(num_voxels);
+
+    const auto actual_vector_multiplier = vector_multiplier_ * soil_model.m_dx;
+    switch (static_cast<SoilProperty>(vector_soil_property_)) {
+        /*
+        case SoilProperty::WaterDensityGradient:
+        {
+                Jobs::ParallelFor(numVoxels, [&](unsigned i)
+                        {
+                                const auto targetVector = glm::vec3(soilModel.m_w_grad_x[i], soilModel.m_w_grad_y[i],
+        soilModel.m_w_grad_z[i]); const auto start =
+        soilModel.GetPositionFromCoordinate(soilModel.GetCoordinateFromIndex(i)); const auto end = start + targetVector
+        * actualVectorMultiplier; const auto direction = glm::normalize(end - start); glm::quat rotation =
+        glm::quatLookAt(direction, glm::vec3(direction.y, direction.z, direction.x)); rotation *=
+        glm::quat(glm::vec3(glm::radians(90.0f), 0.0f, 0.0f)); const auto length = glm::distance(end, start) / 2.0f;
+                                const auto width = glm::min(vector_line_max_width_, length * vector_line_width_factor_);
+                                const auto model = glm::translate((start + end) / 2.0f) * glm::mat4_cast(rotation) *
+                                        glm::scale(glm::vec3(width, length, width));
+                                particleInfos[i] = model;
+                        }, results);
+        }break;*/
+        /*
+        case SoilProperty::Divergence:
+        {
+                Jobs::ParallelFor(numVoxels, [&](unsigned i)
+                        {
+                                const auto targetVector = glm::vec3(soilModel.m_div_diff_x[i],
+        soilModel.m_div_diff_y[i], soilModel.m_div_diff_z[i]); const auto start =
+        soilModel.GetPositionFromCoordinate(soilModel.GetCoordinateFromIndex(i)); const auto end = start + targetVector
+        * actualVectorMultiplier; const auto direction = glm::normalize(end - start); glm::quat rotation =
+        glm::quatLookAt(direction, glm::vec3(direction.y, direction.z, direction.x)); rotation *=
+        glm::quat(glm::vec3(glm::radians(90.0f), 0.0f, 0.0f)); const auto length = glm::distance(end, start) / 2.0f;
+                                const auto width = glm::min(vector_line_max_width_, length * vector_line_width_factor_);
+                                const auto model = glm::translate((start + end) / 2.0f) * glm::mat4_cast(rotation) *
+                                        glm::scale(glm::vec3(width, length, width));
+                                particleInfos[i] = model;
+                        }, results);
+        }break;
+        */
+      default: {
+        Jobs::RunParallelFor(num_voxels, [&](unsigned i) {
+          particle_infos[i].instance_matrix.value =
+              glm::translate(soil_model.GetPositionFromCoordinate(soil_model.GetCoordinateFromIndex(i))) *
+              glm::mat4_cast(glm::quat(glm::vec3(0.0f))) * glm::scale(glm::vec3(0.0f));
+        });
+      } break;
+    }
+    Jobs::RunParallelFor(num_voxels, [&](unsigned i) {
+      particle_infos[i].instance_color = vector_base_color_;
+    });
+
+    ground_fruit_matrices_->SetParticleInfos(particle_infos);
+    update_vector_matrices_ = false;
+  }
+
+  const auto editor_layer = Application::GetLayer<EditorLayer>();
+  GizmoSettings gizmo_settings;
+  gizmo_settings.draw_settings.blending = true;
+  gizmo_settings.draw_settings.blending_src_factor = VK_BLEND_FACTOR_SRC_ALPHA;
+  gizmo_settings.draw_settings.blending_dst_factor = VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA;
+  gizmo_settings.draw_settings.cull_mode = VK_CULL_MODE_BACK_BIT;
+
+  editor_layer->DrawGizmoMeshInstancedColored(Resources::GetResource<Mesh>("PRIMITIVE_CYLINDER"), vector_matrices_,
+                                              glm::mat4(1.0f), 1.0f, gizmo_settings);
+}
+
+
+std::weak_ptr<Soil> EcoSysLabLayer::FindSoil() {
+  const auto scene = Application::GetActiveScene();
+  const std::vector<Entity>* soil_entities = scene->UnsafeGetPrivateComponentOwnersList<Soil>();
+  if (soil_entities && !soil_entities->empty()) {
+    return scene->GetOrSetPrivateComponent<Soil>(soil_entities->at(0));
+  }
+  return {};
+}
