+++ Resources/DigitalAgricultureProject/SorghumField/season12.txt.evefilemeta
@@ -1,4 +1,4 @@
 asset_extension_: .txt
 asset_file_name_: season12
 asset_type_name_: Binary
-asset_handle_: 16676846614897079341
+asset_handle_: 8394492133365956958
