+++ Resources/DigitalAgricultureProject/SorghumField/season12original.txt.evefilemeta
@@ -1,4 +1,4 @@
 asset_extension_: .txt
 asset_file_name_: season12original
 asset_type_name_: Binary
-asset_handle_: 15567102463264743173
+asset_handle_: 6626617199953198205
